requirements.txt
black==23.1.0
isort==5.12.0
pytest==7.2.2
pytest-cov==2.7.1
pytest-timeout==2.1.0
pytest-fail-slow==0.3.0
matplotlib==3.7.1
moviepy==1.0.3
kornia==0.6.12
mypy==0.982
mypy-extensions==0.4.3
tensorboard==2.13.0
opencv-contrib-python==4.7.0.72
numba==0.57.1

# ray
ray[default]==2.3
grpcio==1.48.2 # tested this version with ray 2.3

# simple recon
antialiased-cnns==0.3
efficientnet-pytorch==0.7.1
timm==0.9.2
trimesh==3.22.0
transforms3d==0.4.1
einops==0.6.1
moviepy==1.0.3
pyrender==0.1.45
open3d==0.17.0
scipy==1.10.1
protobuf<4.21.0
setuptools==59.5.0
types-PyYAML==6.0.12.10

# static analysis
pandas==1.4.1
click==8.0.0
tabulate==0.8.7
loguru==0.6.0
darglint==1.8.1

# testing
coverage==6.5

# configs
hydra-core==1.3.2

#planeRCNN data generation
plyfile==1.0.1